@@ -8,12 +8,12 @@ class TimeZone {
   TimeZone._() {
     tz.initializeTimeZones();
   }
-  static TimeZone _this;
+  static TimeZone? _this;
 
   Future<String> getTimeZoneName() async =>
       FlutterNativeTimezone.getLocalTimezone();
 
-  Future<tz.Location> getLocation([String timeZoneName]) async {
+  Future<tz.Location> getLocation([String? timeZoneName]) async {
     if (timeZoneName == null || timeZoneName.isEmpty) {
       timeZoneName = await getTimeZoneName();
     }
