@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 class UserAvatar extends StatelessWidget {
-  final bool ismale;
+  final bool? ismale;
   const UserAvatar({
-    Key key,
+    Key? key,
     this.ismale,
-    @required this.radius,
+    required this.radius,
   }) : super(key: key);
 
   final double radius;
@@ -14,7 +14,7 @@ class UserAvatar extends StatelessWidget {
   Widget build(BuildContext context) {
     return CircleAvatar(
       radius: radius,
-      backgroundImage: ismale
+      backgroundImage: ismale!
           ? AssetImage('assets/images/male-user-1.png')
           : AssetImage('assets/images/female-user-1.png'),
       backgroundColor: Theme.of(context).scaffoldBackgroundColor,
