@@ -14,21 +14,21 @@ class ProfileController extends GetxController {
   String taskBox = 'tasks';
 
   // Text feild controllers for edit account
-  TextEditingController firstNameController;
-  TextEditingController lastNameController;
-  TextEditingController emailController;
+  TextEditingController? firstNameController;
+  TextEditingController? lastNameController;
+  TextEditingController? emailController;
 
   // list of state for gender toggleButton
-  List<bool> selectedToggleGender;
+  late List<bool> selectedToggleGender;
 
   // userData
-  String fName;
-  String lName;
-  String email;
-  bool isMale;
+  String? fName;
+  String? lName;
+  String? email;
+  bool? isMale;
 
   // instance of user model
-  User user;
+  late User user;
 
   //function to toggle gender in edit account
   onToggledGender(int index) {
@@ -48,7 +48,7 @@ class ProfileController extends GetxController {
   }
 
   // FirstName validator
-  String fNameValidator(String value) {
+  String? fNameValidator(String value) {
     if (value.isEmpty) {
       return 'Enter your first name';
     }
@@ -59,7 +59,7 @@ class ProfileController extends GetxController {
   }
 
   // LastName validator
-  String lNameValidator(String value) {
+  String? lNameValidator(String value) {
     if (value.isEmpty) {
       return null;
     }
@@ -70,7 +70,7 @@ class ProfileController extends GetxController {
   }
 
   // Email validator
-  String emailValidator(String value) {
+  String? emailValidator(String value) {
     if (value.isEmpty) {
       return 'Enter your email';
     }
@@ -83,9 +83,9 @@ class ProfileController extends GetxController {
   // function to validate creds
   bool validateCreds() {
     bool validated = false;
-    if (fNameValidator(firstNameController.text) == null &&
-        lNameValidator(lastNameController.text) == null &&
-        emailValidator(emailController.text) == null &&
+    if (fNameValidator(firstNameController!.text) == null &&
+        lNameValidator(lastNameController!.text) == null &&
+        emailValidator(emailController!.text) == null &&
         selectedToggleGender.contains(true)) {
       validated = true;
     } else {
@@ -122,10 +122,10 @@ class ProfileController extends GetxController {
 
   // function to get current userInfo for edit
   preEditUser() {
-    firstNameController.text = fName;
-    lastNameController.text = lName;
-    emailController.text = email;
-    if (isMale) {
+    firstNameController!.text = fName!;
+    lastNameController!.text = lName!;
+    emailController!.text = email!;
+    if (isMale!) {
       selectedToggleGender.first = true;
       selectedToggleGender.last = false;
     } else {
@@ -139,9 +139,9 @@ class ProfileController extends GetxController {
     if (validateCreds()) {
       HomeController c = Get.put(HomeController());
       user = User();
-      user.firstName = firstNameController.text;
-      user.lastName = lastNameController.text;
-      user.email = emailController.text;
+      user.firstName = firstNameController!.text;
+      user.lastName = lastNameController!.text;
+      user.email = emailController!.text;
       user.isMale = selectedToggleGender.first;
       userData.write('fName', user.firstName);
       userData.write('lName', user.lastName);
