@@ -7,15 +7,17 @@ import 'package:taskly/app/modules/settings/controllers/settings_controller.dart
 import 'package:taskly/app/theme/app_theme.dart';
 
 class SplashScreenController extends GetxController
-    with SingleGetTickerProviderMixin {
+    // with SingleGetTickerProviderMixin {
+    with
+        GetSingleTickerProviderStateMixin {
   // animation controller for lottie
-  AnimationController animationController;
+  AnimationController? animationController;
 
   // initializing userData instance
   final userData = GetStorage();
 
   // store Login status
-  bool isLoggedIn;
+  bool? isLoggedIn;
 
   // function to set the user saved theme on start
   setTheme() {
@@ -41,18 +43,18 @@ class SplashScreenController extends GetxController
       duration: Duration(seconds: 3),
       vsync: this,
     );
-    animationController.forward();
-    animationController.addStatusListener((status) {
+    animationController!.forward();
+    animationController!.addStatusListener((status) {
       if (status == AnimationStatus.completed) {
-        animationController.reverse();
+        animationController!.reverse();
       } else if (status == AnimationStatus.dismissed) {
-        animationController.forward();
+        animationController!.forward();
       }
     });
     checkLogin();
     Timer(
       Duration(milliseconds: 5500),
-      () => Get.offNamed(isLoggedIn ? '/home' : '/welcome'),
+      () => Get.offNamed(isLoggedIn! ? '/home' : '/welcome'),
     );
   }
 
@@ -64,7 +66,7 @@ class SplashScreenController extends GetxController
 
   @override
   void onClose() async {
-    animationController.dispose();
+    animationController!.dispose();
     super.onClose();
   }
 }
